+++ docs/changelog.md
@@ -24,6 +24,7 @@ Sprint 12 / May 10, 2020
 - added password validation to Login page.
 - added event prop to form submission to `preventDefault()` if password validation fails on the Login page.
 - updated password validation text to continually display unless the user clicks the Login button.
+- updated auth failure message on Login page to be stored and updateable in state.
 
 <hr />